// frontend/src/app/apiClient/model/companyCase.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



import { BigDecimal } from 'bigdecimal';
import { CasePatientAssessment } from './casePatientAssessment';
import { CasePatientProgress } from './casePatientProgress';
import { CaseEscortReceipt } from './caseEscortReceipt';
import { CaseDocument } from './caseDocument';
import { CaseMessage } from './caseMessage';


export interface CaseEscort {
    escortID: string;
    name?: string;
    email?: string;
    paid?: boolean;
}


export interface EscortLocation {
    escortID: string;
    date: string;
    latitude: number;
    longitude: number;
    stage: string;
}


export interface CaseStatusChange {     
    oldStatus: string;    
    newStatus: string;    
    date: string;
}


export interface CaseTransportConsent {
    signature: string;
    signersName: string;
    signatureDate: string;
    signersRelationshipToPatient: string;
    patientName: string;
    patientDOB: string;
    fromLocation: string;
    toLocation: string;
}


export interface CompanyCase { 
    caseID: string;
    companyID: string;
    companyName?: string;
    caseNumber: string;
    currentStatus: string;

    patientFirstName?: string;
    patientLastName?: string;
    diagnosis?: string;
    firstDayOfTravel?: string;
    numberTravelDays?: BigDecimal;
    originCity?: string;
    destinationCity?: string;
    quotedPrice?: BigDecimal;
    invoiceSent?: boolean;
    invoicePaid?: boolean;
    flightNumber1?: string;
    connectionCity1?: string;
    flightNumber2?: string;
    connectionCity2?: string;
    flightNumber3?: string;
    payPerDay?: BigDecimal;
    externalAccessEmail1?: string;
    externalAccessEmail2?: string;
    externalAccessEmail3?: string;

    patientConsent? : CaseTransportConsent;
    patientAssessment? : CasePatientAssessment;
    patientProgress? : CasePatientProgress;

    escorts: Array<CaseEscort>;
    escortTracking?: Array<EscortLocation>;
    escortReceipts? : Array<CaseEscortReceipt>
    statusChanges: Array<CaseStatusChange>;
    documents : Array<CaseDocument>;
    messages : Array<CaseMessage>;

}